fix(cli): wrap a long empty-result hint with a hanging indent
A hint longer than the terminal hard-wrapped at the edge and the
continuation flushed to column zero, undoing the indent that subordinates
the hint to its message. Both lines of the report now wrap to the width of
the channel they are written to, with continuations aligned under the first
line's own text.

The word wrapper the readiness renderer already used moves into the core
library so both callers share one implementation.
